+++ Sweep/Activity/Pots/Pot.swift
@@ -7,7 +7,7 @@ import Foundation
 public struct Pot: Decodable {
     public let id: String
     public let name: String
-    public let balance: Int
+    public let balance: Int64
     public let currency: String
     public let deleted: Bool
 }
